requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
elasticsearch==7.17.9
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
google-generativeai==0.8.3
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
bertopic==0.16.0
sentence-transformers==2.2.2
